+++ app/src/main/java/kr/ksw/mybudget/presentation/components/SpendingCard.kt
@@ -28,9 +28,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import kr.ksw.mybudget.R
 import kr.ksw.mybudget.domain.model.SpendingItem
-import kr.ksw.mybudget.presentation.common.DATE_FORMAT_YMD_E
-import kr.ksw.mybudget.presentation.common.toDisplayString
-import kr.ksw.mybudget.presentation.common.toPriceString
+import kr.ksw.mybudget.presentation.core.common.DATE_FORMAT_YMD_E
+import kr.ksw.mybudget.presentation.core.common.toDisplayString
+import kr.ksw.mybudget.presentation.core.common.toPriceString
 import kr.ksw.mybudget.presentation.home.spendingList
 import kr.ksw.mybudget.ui.theme.MyBudgetTheme
 import kr.ksw.mybudget.ui.theme.grayTextColor
