+++ app/src/main/java/kr/ksw/mybudget/presentation/core/keys/IntentKeys.kt
@@ -1,0 +1,3 @@
+package kr.ksw.mybudget.presentation.core.keys
+
+const val SPENDING_ITEM_KEY = "spendingItem"
