--- app/src/main/java/kr/ksw/mybudget/presentation/keys/IntentKeys.kt
@@ -1,3 +1,0 @@
-package kr.ksw.mybudget.presentation.keys
-
-const val SPENDING_ITEM_KEY = "spendingItem"
